Adding baseline info to monthly visits
Need to use join of X-1 and Baseline for this task. Right now only X-1
is being used, but there are some incorrect data that is recorded in
X-1.
However, we cannot solely rely on baseline data because the Date of
phone distribution and the date of withdraw are ONLY recorded in X-1.